requirements.txt
hikari==2.0.0.dev111
hikari-lightbulb==2.2.4
pyfiglet==0.8.post1
